refactor: add typing + docstrings, create separate UploadTime component
